requirements.txt
google-generativeai
pandas
openpyxl
Pillow
streamlit
